keep track of BRIGHT in the coadd_exp.mask
This coadding mask is currently only being set for EDGE and if it is set
it is considered a bug and raises an exception.  Also add in BRIGHT and
SAT anywhere BRIGHT was set in the warps.  The SAT will prevent
detection in those areas

Updated interp flags to include BRIGHT as it is now required to be added
to the mask plane by the caller.  No longer setting SAT for BRIGHT
at that stage

added verification that EDGE and NO_DATA are not set in the warps

added verification that EDGE is not in the coadd

corrected some typos